+++ apps/client/app/(dashboard)/portal/[id]/page.tsx
@@ -13,7 +13,6 @@ export default async function CreatePortal({ params }: { params: { id: string } 
     })
   ).data;
 
-  console.log(portal);
   return (
     <div className="my-10 px-3 sm:px-6 md:px-14 lg:px-20">
       <div className="w-full lg:flex gap-4 justify-between">
@@ -41,6 +40,7 @@ export default async function CreatePortal({ params }: { params: { id: string } 
         description={portal.description}
         contributors={portal.contributors}
         updates={portal.updates}
+        owner={portal.user.username}
       />
     </div>
   );
